feat: simplify structure
