+++ payments_py/data_models.py
@@ -1,6 +1,7 @@
 from pydantic import BaseModel, ConfigDict, Field
 from enum import Enum
-from typing import Dict, List, Optional, Union
+from typing import Dict, List, Optional, Union, Any
+from datetime import datetime
 
 class SubscriptionType(str, Enum):
     CREDITS = 'credits'
@@ -95,8 +96,6 @@ class ServiceTokenResultDto(BaseModel):
             }
         })
        
-
-
 class AgentExecutionStatus(Enum):
     Pending = "Pending"
     In_Progress = "In_Progress"
@@ -105,58 +104,63 @@ class AgentExecutionStatus(Enum):
     Failed = "Failed"
     # Add other status options here
 
-# class Artifact(BaseModel):
-#     artifactId: str
-#     url: str
+# Artifact data class
+class Artifact:
+    """
+    Represents an artifact with a unique identifier and a URL reference.
+    """
+    artifact_id: str
+    url: str
 
-# class BaseStepDto(BaseModel):
-#     task_id: str = Field(..., description="Id of the task")
-    
-#     input_query: Optional[str] = Field(None, example="What's the weather in NY now?", description="Input for the task")
-    
-#     input_params: Optional[List[Dict[str, str]]] = Field(None, example=[{'assistantId': '1234'}], description="List of additional key-value parameters required for the step")
-    
-#     input_artifacts: Optional[List[Artifact]] = Field(None, example=[{'artifactId': 'art-aabb', 'url': 'https://nevermined.io/file.txt'}], description="Artifacts for the step")
-    
-#     name: Optional[str] = Field(None, example="summarizer", description="Name of the step")
-    
-#     order: Optional[int] = Field(None, example=1, description="Order of the execution of the step")
-    
-#     cost: Optional[int] = Field(None, example=5, description="Cost in credits of executing the step")
-    
-#     predecessor: Optional[str] = Field(None, example=AgentStepEntity.generateId(), description="Previous step id. If not given, the system will associate to the latest step (by order).")
-    
-#     is_last: Optional[bool] = Field(None, example=True, description="Is the last step of the task?")
+# ExecutionInput data class
+class ExecutionInput:
+    """
+    Represents the input for a task, such as a query and additional parameters or artifacts.
+    """
+    query: str
+    additional_params: Optional[List[Dict[str, str]]] = None
+    artifacts: Optional[List[Artifact]] = None
 
-# class NewStepDto(BaseStepDto):
-#     step_id: Optional[str] = Field(None, description="Id of the step. If not given or invalid, the system will auto-generate it", example=AgentStepEntity.generateId())
-    
-#     step_status: Optional[AgentExecutionStatus] = Field(None, description="Status of the step", example=AgentExecutionStatus.Not_Ready)
-    
-#     input_query: str = Field(..., example="What's the weather in NY now?", description="Input for the task")
+# ExecutionOutput data class
+class ExecutionOutput:
+    """
+    Represents the output of a task or step execution.
+    """
+    output: Any
+    additional_output: Optional[List[Dict[str, Any]]] = None
+    artifacts: Optional[List[str]] = None
 
-# class UpdateStepDto(BaseStepDto):
-#     step_id: str = Field(..., description="Id of the step. If not given or invalid, the system will auto-generate it", example=AgentStepEntity.generateId())
-    
-#     step_status: AgentExecutionStatus = Field(..., description="New status of the step", example=AgentExecutionStatus.Completed)
-    
-#     output: Optional[str] = Field(None, example="success", description="Output of the step")
-    
-#     output_additional: Optional[List[Dict[str, str]]] = Field(None, example=[{'message': 'success'}], description="List of additional key-value output values generated by the step")
-    
-#     output_artifacts: Optional[List[Artifact]] = Field(None, example=[{'artifactId': 'art-aabb', 'url': 'https://nevermined.io/file.txt'}], description="Artifacts generated by the execution of the step")
+# ExecutionOptions data class
+class ExecutionOptions:
+    """
+    Represents options for executing a task or step, such as input, status, and output.
+    """
+    input: ExecutionInput
+    status: AgentExecutionStatus
+    output: Optional[ExecutionOutput] = None
+    created_at: Optional[datetime] = None
+    updated_at: Optional[datetime] = None
+    retries: Optional[int] = None
 
-# class CreateStepsDto(BaseModel):
-#     steps: Optional[List[NewStepDto]] = Field(None, example=[{
-#         'task_id': 'task-1234',
-#         'input_query': 'What is the weather in NY now?',
-#         'name': 'summarizer',
-#         'order': 2,
-#         'is_last': True
-#     }], description="List of new Steps to create")
+# Step data class
+class Step(ExecutionOptions):
+    """
+    Represents a step in the execution of a task.
+    """
+    step_id: str
+    task_id: str
+    is_last: Optional[bool] = False
+    name: Optional[str] = None
 
-# class CreateTaskDto(BaseModel):
-#     query: str = Field(..., example="What's the weather in NY now?", description='Input for the task')
-#     name: Optional[str] = Field(None, example="summarizer", description='Name of the task')
-#     additional_params: Optional[List[Dict[str, str]]] = Field(None, example=[{"assistantId": "1234"}], description='List of additional key-value parameters required for the task')
-#     artifacts: Optional[List[Artifact]] = Field(None, example=[{"artifactId": "art-aabb", "url": "https://nevermined.io/file.txt"}], description='Artifacts for the task')    
+# Task data class
+class Task(ExecutionOptions):
+    """
+    Represents a task that an agent should execute, composed of multiple steps.
+    """
+    task_id: str
+    steps: List[Step]
+    name: Optional[str] = None
+
+# Constants for step names
+FIRST_STEP_NAME = 'init'
+LAST_STEP_NAME = 'init'
